Added ItemCreationController and CollectionItem. Updated Collection. Began Data Entry Integration.
